Quick hack to set an Asset's ID.
Backbone assumes that a model's unique ID is called 'id' and as such uses the existence of this attribute to determine whether an instance is "new" or not, where new means it doesn't exist on the server. However, Screenly's Asset model does not follow this convention, instead using 'asset_id' as the field. I added an initialize method which simply sets the 'id' property to the value of 'asset_id' if it exists.