Fixed: web content black flashes and constant reloads.
Previously, if a playlist with a single web asset was shown it would alternate between showing the web asset and a black screen. This looked like the browser was crashing repeatedly.

Furthermore, even when the black screen was removed the content would be reloaded every time the duration passed, causing pointless visual flashing.

Also previously, the time between web and other content was too long as we paused to clear the browser before loading the new content.

This fix makes it so that we only send UZBL to a new URL if the new URL is actually new. It also postpones the "switch to black" browser clear until after the next bit of content has been loaded. This means the next image is already showing by this time and an ugly black flash is avoided.

Also when switching from web to web content there is no black screen at all in between as the next page is simply loaded. This decreases transition time and looks much better.

Finally, an unoptimised situation where UZBL was loaded with a black screen and then immediately navigated to the same black screen has now been avoided.