Fixed: black screen in case of clock drift + network disconnect.
Without this fix, the last modified time of the asset database could easily be in the future if the RPI forgot the time during a reboot/power off. This would cause viewer.py to constantly regenerate it's playlist, thinking there was a new change, and each time lose track of its play index.

If at the same time the first asset couldn't be played (e.g. web asset with no internet connectivity), a fairly permanent black screen situation would result (at least until the clock caught up with the database mtime).

The fix is to not compare the database mtime with now but instead just check if it increased since the last update.