Ensure that the OOM killer kills the viewer container's init process in an OOM scenario
Currently, the OOM killer opts to kill the renderer processes instead of the init process of the viewer container, resulting in a responsive Pi but with a persistently blank white display. This change ensures that the viewer container's init script is targeted by the OOM killer (in an OOM situation) resulting in a restart of such.

Signed-off-by: Nash Kaminski <nashkaminski@kaminski.io>